# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  description: REST API для управления банковскими картами
  version: 1.0.0
  contact:
    name: Bank Cards API Support
    email: support@bankcards.com

servers:
  - url: http://localhost:8080/api/v1
    description: Development server

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "admin@bankcards.com"
        password:
          type: string
          minLength: 6
          example: "Admin123!"

    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        type:
          type: string
          example: "Bearer"
        expiresIn:
          type: integer
          example: 86400

    CardDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
          example: 1
        number:
          type: string
          example: "**** **** **** 1234"
        holder:
          type: string
          example: "John Doe"
        expirationDate:
          type: string
          format: date
          example: "2025-12-31"
        balance:
          type: number
          format: decimal
          example: 1500.50
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
          example: "ACTIVE"

    TransferRequest:
      type: object
      required:
        - fromCardId
        - toCardId
        - amount
        - cvv
      properties:
        fromCardId:
          type: integer
          format: int64
          example: 1
        toCardId:
          type: integer
          format: int64
          example: 2
        amount:
          type: number
          format: decimal
          minimum: 0.01
          example: 100.00
        cvv:
          type: string
          minLength: 3
          maxLength: 4
          example: "123"
        description:
          type: string
          maxLength: 255
          example: "Transfer between my cards"

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid request parameters"
        path:
          type: string
          example: "/api/v1/cards"

paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Аутентификация пользователя
      description: Получение JWT токенов по email и паролю
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cards:
    get:
      tags:
        - Cards
      summary: Получить список карт
      description: Пользователь видит только свои карты, админ - все карты
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Список карт
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardDto'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer

  /transfers:
    post:
      tags:
        - Transfers
      summary: Выполнить перевод между картами
      description: Перевод средств между своими картами
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Перевод выполнен успешно
        '400':
          description: Ошибка валидации или недостаточно средств
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /transfers/my:
    get:
      tags:
        - Transfers
      summary: История переводов пользователя
      description: Получить историю всех переводов пользователя с пагинацией
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: История переводов
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      type: object
                  totalElements:
                    type: integer